requirements.txt
google-auth
openai~=0.27.2
flask~=2.2.3
flask-sslify
python-dotenv~=1.0.0
